# k-vet-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "k-vet API",
    "description": "Veterinary practice management — single user, self-hosted.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/appointments/{id}/treatments": {
      "get": {
        "tags": [
          "treatments"
        ],
        "operationId": "listTreatments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Treatment"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "treatments"
        ],
        "operationId": "createTreatment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTreatment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Treatment"
                }
              }
            }
          },
          "422": {
            "description": ""
          }
        }
      }
    },
...
    "/api/treatments/{id}": {
      "get": {
        "tags": [
          "treatments"
        ],
        "operationId": "getTreatment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Treatment"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "treatments"
        ],
        "operationId": "deleteTreatment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "treatments"
        ],
        "operationId": "patchTreatment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchTreatment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Treatment"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
    "/api/treatments/{id}/apply-template": {
      "post": {
        "tags": [
          "treatments"
        ],
        "operationId": "applyTemplate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyTemplate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TreatmentItem"
                  }
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
    "/api/treatments/{id}/duplicate": {
      "post": {
        "tags": [
          "treatments"
        ],
        "operationId": "duplicateTreatment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Treatment"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
    "/api/treatments/{id}/items": {
      "get": {
        "tags": [
          "treatments"
        ],
        "operationId": "listTreatmentItems",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TreatmentItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "treatments"
        ],
        "operationId": "createTreatmentItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTreatmentItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreatmentItem"
                }
              }
            }
          },
          "409": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        }
      }
    },
    "/api/treatments/{id}/patients": {
      "post": {
        "tags": [
          "treatments"
        ],
        "operationId": "addTreatmentPatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPatient"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Treatment"
                }
              }
            }
          },
          "422": {
            "description": ""
          }
        }
      }
    },
    "/api/treatments/{id}/patients/{patient_id}": {
      "delete": {
        "tags": [
          "treatments"
        ],
        "operationId": "removeTreatmentPatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Treatment"
                }
              }
            }
          },
          "409": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApplyTemplate": {
        "type": "object",
        "required": [
          "template_id"
        ],
        "properties": {
          "template_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CreateTreatment": {
        "type": "object",
        "properties": {
          "treatment_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "finding": {
            "type": [
              "string",
              "null"
            ]
          },
          "patient_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Patients to attach right away; all must belong to one customer."
          }
        }
      },
      "CreateTreatmentItem": {
        "type": "object",
        "required": [
          "kind"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/TreatmentItemKind"
          },
          "drug_packaging_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Required for `drug_packaging` lines."
          },
          "service_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Required for `service` lines."
          },
          "patient_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Defaults to the treatment's only patient when there is exactly one."
          },
          "quantity": {
            "type": "string"
          },
          "km": {
            "type": [
              "string",
              "null"
            ],
            "description": "Travel-expense lines: kilometres driven (one way)."
          },
          "km_multiplier": {
            "type": [
              "string",
              "null"
            ],
            "description": "Multiplier for adverse travel conditions (1–3)."
          }
        }
      },
...
      "DuplicateRequest": {
        "type": "object",
        "properties": {
          "price_mode": {
            "$ref": "#/components/schemas/PriceMode"
          }
        }
      },
...
      "PatchTreatment": {
        "type": "object",
        "properties": {
          "treatment_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "finding": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Treatment": {
        "type": "object",
        "required": [
          "id",
          "appointment_id",
          "patients",
          "customer_emails",
          "frozen",
          "total_gross",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "appointment_id": {
            "type": "integer",
            "format": "int64"
          },
          "starts_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "treatment_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "finding": {
            "type": [
              "string",
              "null"
            ]
          },
          "patients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TreatmentPatient"
            }
          },
          "customer_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Derived from the patients — the invoice's customer."
          },
          "customer_emails": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The customer's email addresses, offered as invoice recipients (FR-031)."
          },
          "invoice": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TreatmentInvoice",
                "description": "The live invoice of this treatment, or the most recent cancelled one — the vet has\nto see that the last attempt was cancelled, and can then bill again."
              }
            ]
          },
          "frozen": {
            "type": "boolean",
            "description": "`true` once the invoice is accepted: lines and stock movements are frozen."
          },
          "total_gross": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}